+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+
+## [v2.3.0]
 ### Added
 - added StaticValueClaimSource
 ### Changed
@@ -277,6 +279,7 @@ All notable changes to this project will be documented in this file.
 First release of modified MITREid server which uses Shibboleth for authentication and reads user data from Perun system. It also has support for new scopes and claims, for extension of released access tokens, and for releasing user claims from introspection endpoint.
 
 [Unreleased]: https://github.com/CESNET/perun-mitreid/tree/master
+[v2.3.0]: https://github.com/CESNET/perun-mitreid/releases/tag/v2.3.0
 [v2.2.0]: https://github.com/CESNET/perun-mitreid/releases/tag/v2.2.0
 [v2.1.1]: https://github.com/CESNET/perun-mitreid/releases/tag/v2.1.1
 [v2.1.0]: https://github.com/CESNET/perun-mitreid/releases/tag/v2.1.0